myAPI/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-rest==0.8.7
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.3
yarg==0.1.9
